add first widgets of Main Module ALF
Prototype of Alf Main Module View based on the work of hiesl@
and pippel@.
There are parts of the GUI that are supplied with areas called PLACEHOLDERS
and the likes.
As discussed with pippel@ We use them until we have real and fitting widgets
at hand.
Change contains a test trying to make use of the majority of Buttons/Tabs.

Aegis-Change: anoubis.10.650.C426 Delta 54
Bug: 284
BLE