200-200-100 Client und Serverkomponenten - connect gui and daemon
Connect xanoubis with anoubisd.
For this the Communicator was changed to 'be' an own thread, this
shall enable us to use the same logic as in the daemon.
In addition arriving notifies will be handed to ModAnoubis.
The NotifyListElement was extended to extract the informatoion
of an arrived notification.

Aegis-Change: anoubis.10.650.C469 Delta 93
Bug: 300
TE BLE