Collect in kernel status information
The kernel should collect some statistical data and provide it to user space upon request. This change implements statistic
gathering and an ioctl that triggers sending the data as an eventdev event. While we are at it also add proper security checks to
the other two ioctls on the anoubis device and remove some ugly SPLINT macros.

Aegis-Change: anoubis.10.700.C524 Delta 30
Bug: 383