clean all lint reports before running the linters
The reports of the linters currently in the baseline
are plain outdated (and therefore wrong). The issue is most probably that
the reports cannot be integrated properly into the dependency chain of the
autotool project. Therefore this change cleans all linting reports
unconditionally before generating them, so that there can be no leftovers
of previous linting runs.

Aegis-Change: anoubis.10.700.C111 Delta 50
Bug: 377
TE BLE