provide true monotony for pristine source versioning
The versioning of the pristine sources heavily relates on
Aegis' aesub.
We also have to consider non-Aegis environments and therefore have to
provide a static fallback.
Furthermore we introduce a kind of true monotony towards the versioning of
the pristine sources by using a static prefix, base- and branchname plus
the corresponding deltas.
Additionally a new helper-function vsort.pl provided by fritsch@ is used to
assure the correct order of the pristine sources in the stamp-file which is
read by packaging.cook to fetch the latest source.
Attention! This changes needs a mandatory minimal Integration as we changed
the versioning in a direction that leads to newer packages having version
numbers which are less than existing ones.
Additional comments were added as discussed with and recommended by [RT].
This change also appends include_HEADERS = splint-includes.h as we missed
to add the file (as discussed with [RT]) to the pristine source in a former
change.

Aegis-Change: anoubis.10.700.C574 Delta 86
Bug: 163
BLE