fix the behaviour of amodal Dialogs
Add onClose()-handler for the amodal Dialogs LogViewer and
RuleEditor to keep the state of the toggleButtons in sync when the Dialogs
are closed.
Class DlgLogViewerBase now inherits from wxDialog to be consistent with the
behaviour of the RuleEditor.
Furthermore we provide as discussed with hiesl@ onClose()-handler in the
BaseClasses which solves the problem of a segfault which occures when the
toggleButtons are pressed after closing the Dialogs.
Additionally now we use the already existing Events for the Dialogs
LogViewer and RuleEditor and register handlers for them in the
corresponding Views.
This Change modifies the existing LDTP-Tests so that at least toggling of the
amodal Dialogs happens a few times which before ended up with a segfault of
xanoubis

Aegis-Change: anoubis.10.700.C612 Delta 95
Bug: 359
BLE