Implement events to keep TrayIcon and MainFrame informed about connection state and message count
This change introduces new Events which are observed by the
TrayIcon and the MainFrame to keep the status information in sync.
To provide these functionality we have to introduce private members which
are holding the counts of alerts and escalations.
This is especially mandatory to respect the priority of escalations over
alerts and the normal state.
Additionally the TrayIcon now holds all three state icons which also have to
be shown depending on the current state.
Direct calls to update() of the TrayIcon were removed as we are going to
use event-triggered updates (like the ones introduced) in the future.
Eventually we also get rid of a couple of widespread update methods.
This change also adds an additional gui test TrayIconConnectState.pm which,
heavily based on ideas from [SF], does a lookup for the MessageLabel showing
the current state which is expected to be around as we have two LogNotifies
and one AlertNotify after calling connect.
Future changes will also take care of the TrayIcon state.
The change is going to be reviewed by [CH].

Aegis-Change: anoubis.10.700.C620 Delta 100
Bug: 421