implement Alert Event on connection failure
Change the generation (in case of a connection failure to)
of Event to type Alert.
To handle all connection states we introduce a new enum type
connectionStateType {CONNECTION_CONNECTED, CONNECTION_DISCONNECTED,
CONNECTION_FAILED}; which is used by the Communicator and CommunicatorCtrl
to map the connection state cases to the corresponding events.
Additionally the actually shown status message was also adapted.
This Change touches the test TrayIconConnectState.pm which is with these
mandatory adaption back on the map.

Aegis-Change: anoubis.10.700.C642 Delta 113
Bug: 428