implement PopUpMenue on TrayIcon to provide access to xanoubis
This change implements a PopUpMenue to be used with the TrayIcon.
For now we provide basic functionality (close and restore of xanoubis)
nethertheless it can be eventually enhanced to provide also access to the
LogViewer and Ruleeditor and further shortcuts.
We had to introduce new Events in order to accomplish the task of hide and
restore of the MainFrame.
Furthermore the File section of the Menubar now holds a Close option which
also hides the MainFrame.
This change lacks a specific test as there is currently no way to trigger the
context menu via right mouse click. Therefore we only enhance an already
existing Test to also take care of the existence of the TrayIcon itself.

Aegis-Change: anoubis.10.700.C643 Delta 123
Bug: 421
BLE