make notification via libnotify chooseable
In order to provide the user with a choice of getting urgent
alerts and escalation add a basic configuration option to ModAnoubis.
For now we default to not send notifications via libnotify as of the
TrayIcon.
Nevertheless the user can now choose if he wants to receive additional
information via Popups.
The GUI test ModAnoubisNotifies.pm was adapted and enhanced to take care of
the current changes.

Aegis-Change: anoubis.10.700.C652 Delta 126
Bug: 422