Cleanup amsg.c and amsg.h
This is intended to be a cleanups that goes beyond the immediate bugfix from Change 663. Most important changes:
* Move most private declarations from amsg.h to amsg.c. This avoids
redclaration of the fds-Array in each file that includes amsg.h
* Remove event pointer from struct msg_buf. It is only used in a single
error path where it is not really needed.
* Fix handling of pending write events if the message can be written to
the buffer but not to the file descriptor.

Aegis-Change: anoubis.10.700.C670 Delta 132
Bug: 439
TE BLE