GUI policy handling - first steps
This change adds first pieces of policy handling
to xanoubis. For now, this is a menu entry in the file section
of MainFrame. By selecting it, a file dialog will been opened.
After a (policy) file was choosen, the apn_parse() from libapn
is called. The resulting data structure is propagated via AnEvent
to ModAlf, ModSfs and RuleEditor to be displayed.

Aegis-Change: anoubis.10.700.C625 Delta 163
Bug: 413
TE BLE