fix the rendering of Messages in ModAnoubis
So far only the growableCol seems to work (at least for the
rendering).
Trying to add an additional spacer and also spurious static text screwed
the layout further.
Therefore we stay with this interims solution as discussed with [CH].
This change only touches the corresponding test to assure that the current
state doesn't change it's behaviour.

Aegis-Change: anoubis.10.700.C713 Delta 166
Bug: 451
BLE