spring-loaded GUI fixes
This change was discussed with [CH] and takes care of several
last minute fixes regarding xanoubis.
This change fixes (a quick hack) the problem of the gui to abort on
ack of escalations.
Furthermore it enables the rendering of all notification types.
Additionally it fixes the update of MainFrame and TrayIcon regarding
received escalations by respecting the notification priority.

Aegis-Change: anoubis.10.700.C736 Delta 190
Bug: 155
BLE