Replies are not revmoved from replyq
Ask events sent to the UI are put on a replyq for time out handling. If the GUI exits, disconnects or answers pending request to
the daemon. However, the daemon does not remove the answered events from the replyq. Thus, after a time out the daemon tries to
cancel already answered events which causes a write error in _flush_buf() to /dev/eventdev (EINVAL) as the cookie is not valid
anymore.

Aegis-Change: anoubis.10.700.C754 Delta 195
Bug: 463
BLE