GUI policy handling - insert new rules
=== CLONE of 10.700.C684 ===
This change adds first pieces of policy handling
to xanoubis. For now, this is a menu entry in the file section
of MainFrame. By selecting it, a file dialog will been opened.
After a (policy) file was choosen, the apn_parse() from libapn
is called. The resulting data structure is propagated via AnEvent
to ModAlf and RuleEditor to be displayed (which is not completely
done by this change).

Aegis-Change: anoubis.10.700.C760 Delta 198
Bug: 413
TE BLE