Fix two crashes and a typo
xanoubis crashes if it finds a application rule with an empty application ('any'). Additionally there is the crash described in
Bug 481 where wxTreeCtrl->DeleteAllItems() tries to delete the wxTreeItemData object associated with a tree item. However, this
object is not really of type wxTreeItemData, it is just statically cast from AppPolicy/AlfPolicy to that type.

Aegis-Change: anoubis.10.750.C114 Delta 9
Bug: 481
TE BLE