add .desktop file of xanoubis to Debian Package
To appear at least on app menus of Gnome/KDE on Debian add
a basic desktop and icon file.
The icon file is just used for prove of concept and has to be replaced with
a decent one which is hopefully provided by [MPI].
The test TarballComplete.pm was copied to ensure that we don't break that
specific test by the introduces changes.

Aegis-Change: anoubis.10.750.C125 Delta 21
Bug: 459
BLE