fix make target 'make distcheck'
The make target 'distcheck' fails due to failing to find the
shared libraries in the search path.
We add 'TESTS_ENVIRONMENT' to cope with that problem so that shared
libraries are found and therefore the 'make distcheck' works as expected.
Two tests (test/anoubisd and test/xanoubis) are now disabled per default as
both rely on a running anoubisd which can't be expected when running
'make distcheck'.
Nevertheless we still want to ship these tests as part of the pristine source
as both are still quite useful.
This Change adds a new Test 'TarballMakeDistCheck.pm' which tests the ordinary
process of running 'autogen.sh', 'configure' and the 'make distcheck'.

Aegis-Change: anoubis.10.750.C509 Delta 35
Bug: 423