Rule reload can cause illegal context switch
Reloading rules will search the new ruleset for a running process according to the processes pathhint and checksum. This is wrong
if the process inherited its context from a parent process. Note that the parent process cannot be used either because the parent
might have performed a context switch in the mean time. The solution implemented in this change is to use the checksum and path
associated with the currently active ruleset of the process.

Aegis-Change: anoubis.10.750.C554 Delta 73
Bug: 529
BLE