store user specific settings in filesystem and enforce options view on menubar->preferences.
This change provides xanoubis with the ability to store and
restore user settings made in the Options Tab of xanoubis.
Furthermore this changes adds a new Event 'anEVT_ANOUBISOPTIONS_SHOW' to
xanoubis which is used to select the options view of Module Anoubis when the
user clicks Edit->Preferences in the Menubar.

Aegis-Change: anoubis.10.760.C612 Delta 27
Bug: 578
TE BLE