add event handler selecting the corresponding
View of the Ruleeditor when Escalation with depending RuleId was selected
in the Log Viewer.

It's now possible to select entries in the Dialog Log Viewer
triggering the Dialog Rule Editor View with the corresponding RuleId
selected.
The depending Tests DlgLogViewer.pm and DlgRuleEditor.pm were copied to
ensure that nothing breaks.

Aegis-Change: anoubis.10.760.C621 Delta 35
Bug: 506
BLE