incorporate fixes to 'ModAnoubisMainPanelImpl.cpp and 'DlgLogViewer.cpp'.
Change DlgLogViewer::OnListItemSelected() to send an show
event only if the actual event is greater than '0'.
Due to the fact that events, that have no corresponding RuleId return '0'
and therefore are of no interest when it comes to the automatic selection of
the Ruleeditor view.
ModAnoubisMainPanelImpl::OnEscalationsShow() has to call update in order to
sync the view with actual chosen Alerts or Escalations.

Aegis-Change: anoubis.10.760.C631 Delta 39
Bug: 156
BLE