adapt OpenBSD 4.2 packages and PLIST due to new shipped binaries.
This change syncs the package infrastructure for OpenBSD 4.2
and reverts a premature change to packaging.cook
The PLISTs were adapted to create the mandatory device nodes as workaround
until we have 'make release' in place.
The shipped profiles are also copied to the place in the FS were xanoubis
expects them acting as default policies.
This is an interims solution and has to be adapted as soon as possible.
As we are still in the reworking phase of milestone VII we still have to
ship 4.2 packages.

Aegis-Change: anoubis.10.760.C684 Delta 77
Bug: 14
TE BLE