move the creation of event dev and anoubis dev to kernel package.
As discussed with [CEH] it's more appropriate to have the
creation of the event dev and anoubis dev bound to the fake kernel package.
Even though the anoubis-kernel is a fake port we add the port sources to
the tree just to keep the record.

Aegis-Change: anoubis.10.770.C126 Delta 8
Bug: 14
TE BLE