Navigating with ''Tab''
The way i expect the navigation with the tab key works can be found as example in the GUI at the Anoubis module and then under
options. (using the tab key will jump to the next field for edeting)
That does not work that nice in the ruleeditor which would be nice.

Aegis-Change: anoubis.10.800.C150 Delta 7
Bug: 668
BLE