add optional lcov-/gcov-configure Argument to pristine source to support the generation of code coverage reports
via unit tests.

In order to get at least the surplus of code coverage reports
in interaction with the already existing check Unittests we introduce
additional configure options to the pristine sources.
There will be an explicit make-Target 'make lcov' which triggers and
generates the report. Which can later be used to implement a periodical
report generation.

Aegis-Change: anoubis.10.820.C100 Delta 2
Bug: 311
BLE