Implement an SFS exception for connected user Interfaces.
Implement an a posibility for User Interfaces to disalbe SFS rules for the UI itself. This is neccessary for offline checksum
management. While at it fix a bug where we did not translate POLICY_* reply codes to errno numbers in replies. Finally also
implement a checksum request that adds a given checksum to the shadow tree and deprecate request that require the anoubisd to
calculate a checksum.

Aegis-Change: anoubis.10.900.C138 Delta 37
Bug: 771
TE BLE