Remove security check for ADDSUM
As of now the daemon only allows a user to manipulate its own checksum namespace. The daemon never actually accesses the file
referenced in a checksum request. Thus there is no need for a security check anymore. This eliminates the use of switch_uid and
prevents DOS attacks on the daemon by adding checksum for large files.

Aegis-Change: anoubis.10.900.C168 Delta 41
Bug: 771
BLE