Admin policies not visible/editable
This change adds the new class ProfileCtrl controlling the ProfileMgr.
In addition the class PolicyRuleSet is extended by a field for user id
the concerning rule set belongs to. A spell-check was done for ProfileMgr.
While adding the tests for ProfileCtrl to the suite_Profile, the errors
caused by 'libGUIcomponents_DEPENDENCIES' were solved by renameing it
(see Bug #790).

Aegis-Change: anoubis.10.910.C125 Delta 14
Bug: 732