Guess application skript is not working
The script for guessing the application of a given
package or programm is too slow and thus not activated.
As we have no resources left to implement a faster solution
we drop this feature yet.
This change removes all concerning widgets from the gui. The guess-script
itself is not removed, because we may need it in the future.

Aegis-Change: anoubis.10.910.C139 Delta 51
Bug: 609
TE BLE