add @exec @unexec to PLIST of OpenBSD to cope with none system installation of anoubisd(8) package
we have to use a workaround in PLIST of the OpenBSD package
of anoubisd(8) to cope with reinstall, upgrade of package not done via the
system installation (think of the ISO).
In the near future we should eventually get rid of the anoubisd(8) package
and bring the daemon sources into the Anoubis branch of the OpenBSD source
tree.

Aegis-Change: anoubis.10.910.C115 Delta 58
Bug: 14
TE BLE