Rule editing with ''Always allow'' fails
EscalationNotify is creating an anEVT_ANSWER_ESCALATION-event,
which holds the answer of the escalation. Due to the
event-broadcasting of AnoubisGuiApp the answer is scheduled two
times for sending back to anoubisd. Both messages are sent but
the second one is rejected by anoubisd. The
'rejection-notification' send back to the client was not
processed and leads to an error-message.

Aegis-Change: anoubis.10.910.C218 Delta 88
Bug: 865
BLE