'test/libanoubisui/apnvm' breaks code coverage
The test test/libanoubisui/apnvm missed the GCOV_ENABLED
conditional for the LDADD and LDFLAGS which always ended up with trying
to link to the share object of needed libraries. That is still not
supported and therefore one has always link to the static version.

Aegis-Change: anoubis.10.950.C148 Delta 16
Bug: 917
BLE