Inconsistency menu  Tools->{Rule Editor,Log Viewer}
This change fixes the handling of events for the GUI.
A central event table is established and all callback methods
are registered there. Those methodes are extended by a call of
'event.Skip()' if not present already.
A pice of documentation was added to AnEvents.h.

Aegis-Change: anoubis.10.950.C127 Delta 18
Bug: 853
TE BLE