Check for memory leaks
Create test to check parser with some broken policies. Use valgrind to detect the leaks.
Switch to bison to be able to use %destructor to fix the leaks.
The fixes in parse.y are not complete but the minimum to allow to proceed with checks with fuzzed policies.

Aegis-Change: anoubis.10.950.C191 Delta 57
Bug: 409