creating autostart directory fails
The creation of the autostart directories for Gnome and KDE
is likely to fail if the parent directory doesn't exist already.
Therefore switch the creation of the directories to use wxFileName::Mkdir()
and the optional flag 'wxPATH_MKDIR_FULL' in the veins of 'mkdir -p'.

Aegis-Change: anoubis.10.950.C244 Delta 70
Bug: 962
TE BLE