Reimplement policy class structure
This change implements a new types and structure of
policy classes. In addition unit tests were implemented for all
policies. This change is one of a series of changes thus not
complete in all details (e.g. unused code will be removed in the
next change and additional functionality added when needed).

Aegis-Change: anoubis.10.950.C126 Delta 96
Bug: 756