Details widget
This change creates an own directory for special widgets
created for xanoubis. The already existing, like AnStatusBar, are moved
to it. And a new widget for showing details is created.
The new widget will been used within the next change of ModSfs by rdoer@.

Aegis-Change: anoubis.10.950.C334 Delta 137
Bug: 785
TE BLE