anoubisctl reload should be functional
Currently the anoubisctl subcommand ''reload'' is a no-op. The subcommand should be implemented, documented and used in the
initscripts. That way the functionality will be used and tested. And we'll have the option of changing the way we implement
reload functionality at will (adding stuff like signature based reloads)

Aegis-Change: anoubis.10.950.C458 Delta 202
Bug: 1066