Make ruleset reload from Daemon more robust
Currently we always generate an empty ruleset if we generate a load ruleset event. This is not a good idea in general, we should
only do this if no appropriate ruleset exists. While at it, also make the rule editor smarter wrt ruleset reloads.

Aegis-Change: anoubis.10.960.C207 Delta 57
TE BLE