Add missing yyrestart calls to lexers
The yylex() function generated by lex/flex is designed such that it can be called repeatedly on the same input stream. Each call
will return the next token. Thus special care must be taken if the input file of a lexer is modified by assigning to yyinput. Add
this to the lexers in our tree. Additionally set the start state back to INITIAL.

Aegis-Change: anoubis.10.960.C225 Delta 68
Bug: 1145
BLE