Policy load fails if scope is present
After a daemon restart policy parsing fails if a scope is present. Additionally, the daemon can segfault because pe_policy_get
tries to clean a ruleset that is actively used by other parts of the daemon. In particular the rules that are deleted are still
hashed in the policy hash.

Aegis-Change: anoubis.10.970.C194 Delta 53
Bug: 1205