Error handling in case of syntax errors is broken
If apn_print_ruleset returns an error the ruleset is not exported but there is no indication why. Additionally, upon save with
syntax errors an empty policy is sent to the daemon. We should not send the policy at all and give a reasonable error message
instead.

Aegis-Change: anoubis.10.970.C287 Delta 103
Bug: 1106
BLE