Checksum for xanouis is not calculated in Profil after install on FC8
The Profiles with rules for anoubisctl and xanoubis including Checksums. The checksum of anoubisctl is calculated correct but of
xanoubis is missing. Also Fix an issue with the selected page in case of remove and kill a fragile event handling construction
that assumes that DlgRuleEditorAppPage::onDeleteButton is called before AnPolicyNotebook::onDeleteButton.

Aegis-Change: anoubis.10.970.C317 Delta 119
Bug: 1248
BLE