Replace apncvs_makecmd and system with fork/exec
Replace the use of system im apncvs with a fork/exec based solution. This solves problems with escape sequences etc. In a second
step we will be able to get access to the pid of the cvs process and exempt this process from sandbox/sfs events. NB: If it turns
out that cvs spawns additional child processes we can also put it into its own session.

Aegis-Change: anoubis.10.970.C343 Delta 140
BLE