use -fno-working-directory during deb package build to optimize ccache usage
The information about the working dir is different for each version number and prevents ccache hits. This info is only used by
gdb to find the source, but the source is not present on the system where the packages are installed, anyway.
Also fix two misuses of CPPFLAGS.

Aegis-Change: anoubis.10.970.C371 Delta 153
TE BLE