Handling of SFS action and log radio button
These Radio button groups simply translate the index of the selected radio
button into an action or a log number. This accidentally works right now but in
the future we might change the actions from an enum starting at 0 to defines or
change the order of the enum values. Such a change should not break the GUI.
	 - src/anoubis/src/xanoubis/model/SfsFilterPolicy.cpp
	 added new methods, which set the Action and Log-Level
	 depend on the name/string-value of the checkbox
	 - src/anoubis/src/xanoubis/model/SfsFilterPolicy.h
	 Prototypes for new methods
	 - src/anoubis/src/xanoubis/view/DlgRuleEditorFilterSfsPage.cpp
	 Actions now use the new methods and commit the string of the selected
	 RadioButton.
	 - test/gui/unittest/suite_Sfs.pm & test/gui/xanoubis/ViewModSfsGeneric.pm
	 copied Tests to prove/test that nothing has been broken

Aegis-Change: anoubis.10.970.C404 Delta 181
Bug: 1085
BLE