Fix suite_Notifications
Due to missing wxWidgets-infrastructure in tc_Notifications the test fails with an segmentation fault (test sends an event in the
background). Furthermore NotificationCtrl is now using ProcessEvent() for sending the event, if you are in the main-thread. In
the test-environment everything is running in a single thread, thus ProcessEvent() makes the life easier because you don't need
to push the event-loop.

Aegis-Change: anoubis.10.970.C491 Delta 214