refactoring of 'PLIST-daemon' and 'PLIST-main'
Some minor cleanup of 'PLIST-daemon' and 'PLIST-main'.
Add '@bin marker' to 'bin/xanoubis' in 'PLIST-main' and also add
'man/man3/anoubis_ui.3' to 'PLIST-daemon' as it was missing til now.
Additionaly refer to the Anoubis homepage as it's already around and add
a todo to introduce 'MASTER_SITES' for a future place providing the
distfile for package build from source.

Aegis-Change: anoubis.10.970.C502 Delta 219
Bug: 1275
TE BLE