change /dev/anoubis permissions in postinstall
When installing anoubisd for the first time, and the anoubis kernel has been booted before anoubisd is installed, it can happen
that the /dev/anoubis device is created before the udev rules are in place. In this case /dev/anoubis gets the wrong permissions
(640 instead of 644).
Work around this by adjusting the permissions in anoubisd postinst

Aegis-Change: anoubis.10.970.C561 Delta 240
Bug: 1299
TE BLE