add example anoubisd.conf as package-specific infrastructure file
Add example anoubisd.conf as package specific infrastructure
file and install it as part of the binary packages.
So we can later on adjust them to our needs.
Right now the infrastructure files are just stubs.
This changes aims to at least provide a configuration file for the anoubisd
within every binary package we support so far.
The OpenBSD package adaption needs some more time and is going to be
introduced by C589.

Aegis-Change: anoubis.10.970.C584 Delta 259
Bug: 14
TE BLE