Message flushing Bugs
We must try to flush the message buffer in a write event handler even if there is no new message in the outgoing queue. Fix this
in all handlers by calling send_msg unconditionally. Change send_msg so that it just flushes the buffer if the message is a NULL
pointer. In fact this makes the message handling code simpler. Additionally, implement mesage flushing for the upgrade-to-master
queue.

Aegis-Change: anoubis.10.970.C628 Delta 280
BLE