Fix task tracking *sigh*
Add additional tracking function for clone (aka fork) and exit. In the daemon acount for these with a thread count inside struct
pe_proc. Report upgrade end once this count reaches zero. The credentials tracking and the logic for storing path names in the
credentials is not changed.

Aegis-Change: anoubis.10.970.C693 Delta 310