sfssig prints Invaild argument while export
The error is correct but not intressting for a user.
In the beginning we try if the argurment is just a file and not a directory
If it doesn't work we proceed. The error occurs with / as argument. Since
This is a directory we can hide the error message.

Aegis-Change: anoubis.10.970.C710 Delta 316