Improve segfault handling
Currently we block all segfault signals which causes anoubisd to loop
when a segfault occurs. In most cases this makes the segfault extremely
hard to debug because the system will freeze. We should have
a more informative segfault handler which prints enough information
to debug the problem and terminate the daemon.

Aegis-Change: anoubis.10.970.C700 Delta 319
Bug: 1337