regtest pkg-add
This change implements a test to check the upgrade procedure
via pkg-add. Currently the OpenBSD kernel does not sent lock-events of a
directory to the daemon. Thus we need to track the whole process.
The config file of the daemon is adjusted, for packaging too.

Aegis-Change: anoubis.10.970.C703 Delta 324
BLE