Terminate Children if master dies
Make sure that the children properly terminate if the master process dies or otherwise signals EOF on the m2* pipes. If the
master is signalled (e.g. with SIGINT) it will tell the children to terminate by sending signals on its own. However, if the
master just dies (e.g. in case of a SEGFAULT) it will not send these signals and the chidlren will not terminate. Fix this by
simulating a signal in case of an EOF on one of the m2* pipes.

Aegis-Change: anoubis.10.970.C734 Delta 330