FIX: tar content query in install.sub and the helper script 'touch_upgraded.pl'
Use the same mechanism to pipe the query of the
install sets via ftp and read them from the pipe instead of
assuming that the sets exist on the filesystem which is plainly
wrong.
The helper script 'touch_upgraded.pl' must not bailout if '$sfsdir' doesn't
exist in the first place.

Aegis-Change: anoubis.10.971.C103 Delta 1
TE BLE