Make Upgrade-MsgBox configurable
Introduce new Option 'Options/ShowUpgradeMessage' to xanoubis.conf.
going to be used for the upgrade Dialogue being shown when an upgrade happened
and the user wants to be informed about that.
Defaults to 'true'
Read and write the option setting from/to xanoubis.conf.

Aegis-Change: anoubis.10.971.C125 Delta 14
TE BLE