SFS-Cleanup: New implementation of list operations
This is a first shot at a new implementation of the various sfs list operations. The code should ultimately move to sfs.c and
what currently is in localflags should come from user space. This will then replace all the list operations with a single one.
Additionally, make sure that the more restrictive default user policy in the sfs test does not apply to tests run as root.

Aegis-Change: anoubis.10.971.C131 Delta 23
BLE