Add an index file for upgrade markers
This changes adds an index file for upgrade markers. This allows us to iterate effectively over upgraded files and check for
their existance. Note that the presence of an index file is only a hint that there may be upgraded files. Any UI must be able to
handle the case where an upgrade list is non empty but contains only directories.

Aegis-Change: anoubis.10.971.C191 Delta 45
BLE