FIX: '/etc/rc' of OpenBSD source and Ports Makefile
Fix the path of 'upgrade_sfs.pl' in '/etc/rc' and set
'WRKOBJDIR =' in the Ports Makefile to let the OBJDIR reside below the
ports directory.
This should improve the ability of parallel builds as the Package is
already build in a temporary directory.

Aegis-Change: anoubis.10.971.C378 Delta 132
TE BLE