Rule editor: Support App and ctx-filter in FilterSubjectPage
Unify some code and support app and ctx-filter in FilterSubjectPage. These two policies now have an additional binary index that
points to one app in the application. There is already a follow up change that uses this code to make the rule editor aware of
the new application and ctx-filter policies. However, this change kills verification of checksum in the rule editor and needs
some more discussion.

Aegis-Change: anoubis.10.971.C380 Delta 136
TE BLE