Rule editor: Support Subjects in app and ctx-filter
This change supports an almost complete set of subjects in app and ctx-filter rules. However, at the same time it kills the
automatic checksum calculation for these rules. We need to figure out if this is acceptable or how to replace the checksum
calculation.

Aegis-Change: anoubis.10.971.C340 Delta 138
TE BLE