Rule editor: Fix creation of user application policy
If a non-root user tries to create a new application policy while an admin policy is selected, he gets an error message, that the
admin ruleset cannot be modified. This will be fixed: if a non-root user tries to create a new application policy while an admin
policy is selected, a new policy is created.

Aegis-Change: anoubis.10.971.C385 Delta 141
TE BLE