Support display of ruleset that contain borrow
Support proper display of context borrow rules in the rule editor. These kinds of rules can still not be created in the rule
editor. This is on purpose because the feature is only used in special rule files provided for KDE with the packages.

Aegis-Change: anoubis.10.971.C445 Delta 169
TE BLE