Support multiple syntax versions
This change adds support for multiple versions of parse.y that can be used by the same apnparser.c. We always try the most recent
parser first. If that parser fails but succeeded in parsing the version number we try the parser that corresponds to this version
number. This change also unifies code for apn_parse and apn_parse_iovec in apnparser.c.

Aegis-Change: anoubis.10.971.C495 Delta 206
TE BLE