Always add quotes when printing path names
The behaviour with regard to printing of path names and key IDs is inconsistent. Make sure that these are always printed with
quotes. For pathnames this is important because of spaces and for key IDs it is important because key IDs consisting only of
digits will not be parsed as a string without surrounding quotes. Regression tests passed.

Aegis-Change: anoubis.10.971.C513 Delta 212
BLE