fix GUI package dependencies
Change the dependencies of xanoubis to require a minimal
version of anoubisd by adapting the infrastructure files used for
the actual build target.
For OpenBSD we have to add some kind of nasty hack as we build
in a temporary directory below mystuff.
Therefore we have to supply the search directory the pkg tools
use to qualify the dependcy in question.

Aegis-Change: anoubis.10.971.C580 Delta 256
Bug: 1359
TE BLE