FIX: xanoubis log messages should be more consistent
The log messages produced by the GUI (via message dialogues
and the logging mechanism used by the log viewer) are not that consistent.
Therefore messages propagated by message dialogues should be in sync with
the related ones in the log viewer to give the user/admin a chance to
get a clue what actually is going on/happened.
As discussed with valentin@ we get rid of passing version mismatch
messages to the StatusBar as it's redundant information.
o change log message to Debug::info instead of Debug::err.
o remove propagation of logMessage to AnStatusBar
o enable project URL in about box
o add srinkes@ to the developers in about box
o fix a whitespace in 'xanoubis/main.cpp'

Aegis-Change: anoubis.10.971.C586 Delta 265
BLE