Fix: online installation/upgrade via OpenBSD packages lacks copying of related binaries
The ui binaries sfssig, anoubisctl as well as the anoubisd
have to be automatically updated or installed in '/sbin'.
Therefore register them via '@sample'.
This is also true for anoubis-keygen but should reside at '/usr/bin'.
Taken that approach we can get rid of the nasty line in the install.site and
upgrade.site and the one thing left for the user is to restart the daemon.

Aegis-Change: anoubis.10.971.C649 Delta 292
TE BLE