Remove INotify Logic
Unfortunately, INotify is apparently buggy. It generates bogus events for the very common case where we run /sbin/grub-update.
The consequence is that an infinite sequence of Upgrade-Notify Events is generated. Fortunately, we no longer need INotify
because we have an upgrade end event. Thus we only check for changes to the menu.lst as a response to one of these events and
after the initial start of xanoubis.

Aegis-Change: anoubis.10.971.C656 Delta 297
TE BLE