Add a generic IndexTree Pattern and use in Sfs
The new template class supports a sorted collection of objects the can be accessed either based on the sort key or based on their
current index in the collection. Inserting, Removing, key to index and index to object can all be done in logaritmic time. Using
this is the Sfs list saved about 50% CPU time when listing ~100.000 files in /use recursively.

Aegis-Change: anoubis.10.971.C740 Delta 356
BLE