libanoubisprotocol: Authentication
Start implementation of different authentication methods in libanoubisprotocol. This change removes the previously unused
username field from the AUTH_REPLY message and adds hooks in the client and server side that can be used to implement
authentication. Existing transport based authentication is still intact and we do not change the protocol version. An old client
should be perfectly able to connect to a newer daemon (provided that the daemon does not insist on a key based authentication).
New clients should simply try to fall back to the old transport based authentication if connecting with the key based
authentication fails. Regression tests PASSed.

Aegis-Change: anoubis.10.971.C814 Delta 413
Bug: 1399
TE BLE