Fix OpenBSD package build because of moved yacc output file date.c
The openbsd package Makefile removes the pre-generated data.c parser from the extracted anoubis sources, to allow a new version
to be generated. Due to the libanoubis changes the date.c file has moved to a new location, which causes the OpenBSD package
build to fail. However due to upgrades on dev the generated date.c file now works on OpenBSD so the whole operation is obsolete
and should be removed.

Aegis-Change: anoubis.10.972.C283 Delta 100
Bug: 1411
TE BLE