Default policies after a new install should contain context new any
After a fresh install there is no matching policy for a
user as the uid is mostly unknown during the package install.
Nevertheless we have to install a policy default file for users
which must contain 'context new any' that can be picked up on
initial creation of user rules.

Aegis-Change: anoubis.10.972.C310 Delta 117
Bug: 1420
TE BLE