SFS-Browser: Unregister bug
fileExists has only two callers. Both of them are better of they only check for existance and not for the file type. In
particular, the current code behaves strange if a file list contains stuff like devices: If a (potentially non existant) checksum
for that file is removed, the entry vanishes from the file list.

Aegis-Change: anoubis.10.972.C328 Delta 137
TE BLE