Make .xanoubis permssions 0700
Late mkdir in apncvs_init should also use 0700 instead of 0750. This mkdir is a problem, if the GUI fails to start due to a
missing DISPLAY. In this case the _De_structor of AnoubisGuiApp will still initialize the version control which in turn will call
this function.

Aegis-Change: anoubis.10.972.C339 Delta 146
TE BLE