Private key loaded on startup depending on path!
On the startup of xanoubis sometimes the private key is loaded and sometimes not.
After searching the code it became clear that this behaviour depends on the path to the private key which is configured in the ~/
.xanoubis/xanoubis.conf file.
If this path is the same as the default path which is defined in the model/PrivKey class, then the key is NEVER loaded on
startup.
If the configured path differs from this default path, then the key is ALWAYS loaded on startup.
This can't be good, because we don't want the key to be loaded into memory if we don't need it there. Thus we have to make sure,
that the key is only loaded when it's needed independent of it's path in the filesystem.
Reviewer: Christian Hiesl

Aegis-Change: anoubis.10.972.C469 Delta 220
Bug: 1434
TE BLE