Fix memory leak in libapn
If a ruleset is destroyed, the destructors for the rule blocks are not called because the ruleset is passed as NULL from the
toplevel apn_free_chain. However, if the ruleset that is passed is not NULL, we will remove each rule individually from the
rbtree. This adds overhead that should have been avoided by the NULL ruleset. We now use rs->idtree == NULL for this check
instead.

Aegis-Change: anoubis.10.972.C529 Delta 259
BLE