Fix RuleWizard related crash
The RuleWizardHistory object is used by the rule wizard pages. Thus it must survive until all of them are destroy (their
destructors will try to disconnect from history objects). Add a reference count to the rule wizard history and only destory it
after the last reference is gone. This might fix Bug 1462. However, it is not entirely clear if there are other cases similar to
this one, in particular during exit of xanoubis.

Aegis-Change: anoubis.10.972.C536 Delta 263
Bug: 1462
TE BLE