Make sure that private key remains in memory
All tasks that need the private key should take a reference to the key and the key must not go away as long as any of these task
still holds a reference. While at it, get rid of the getInstance() method of KeyCtrl and use explicit instantiation of the base
class method instead. Run the timer twice if the key data was accessed since the last start of the timer.

Aegis-Change: anoubis.10.972.C540 Delta 271
BLE