Start xanoubis minized if in autostart mode
If autostart is set, we start xanoubis minized. This matches the behaviour that most users would expect from an autostarted tray-
icon application. However, this also means that currently, xanoubis will start minimized if the autostart flag is on, even it the
current instance is not autostarted.

Aegis-Change: anoubis.10.972.C547 Delta 272
Bug: 1459
TE BLE