TrayIcon skaliert nicht richtig
Because of more icons been used now, this change
add the usage of AnIconList to the TrayIcon.
In addition we register for size-events and determine the orientation
of the task bar.
Furthermore depending on the size of the task bar, a matching icon-size
is calculated and used to fetch a matching icon.

Aegis-Change: anoubis.10.972.C559 Delta 302
Bug: 828
TE BLE